Change prices from type float to BigDecimal

+++ backend/backend/src/main/java/de/morninggrind/backend/domain/Product.java
@@ -1,13 +1,15 @@
 package de.morninggrind.backend.domain;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 public class Product {
     private UUID id;
     private String name;
-    private float price;
+    private BigDecimal price;
     private boolean isActive;
     private boolean isOffer;
     private Long categoryId;
